Add bootloader and kernel shims (#5)
* Bootloader and kernel shims added

This boot path is selected as the default.  The changes are a shim and
amount to no change from the previous boot flow (where main() was called
directly from SystemInit).

* Moved clock init and ftab init into boot/kernel

* Make weakref to main for standalone builds

The kernel can now be compild without any application to launch. This is
mostly useful for removing the ever-confusing "undefined reference to
main" linker error.