Scheduler nearly complete and refactored.
Unfortunately, in order to do this neatly, a great deal of kernel
structure needed to be added.  Some of it is unnecessary, but worthwhile
in the long run, and ultimately makes the code more segmented and
structured.

 + The kernel now has "ports" - there is only the arm/cm4 port,
 currently.
 + Generic algorithms have their own files/modules at the kernel level,
 but their entry points are defined in the port.
 + The NVIC and Core priorities have been figured out.  Bug reports
 coming in hot to move everything to use NVIC_SetPriority due to
 MSB-aligned implementation (thanks, ARM.)

 - Kernel memory is a hot mess. There is a single file which allocates
 the process table (but not the file table, too many changes at once)
 - The drivers still allocate their own memory, which makes the previous
 bullet a design flaw.

More to come soon.  Switching on the SysTick is up next.  Ideally,
figuring out how to ensure SVC/PendSV do not preempt interrupts would be
nice.