Process switcher test.
Runs three applications in a round-robin scheduler, optionally with
cooperative-only multitasking.  The placement of cooperative
multitasking yield() calls is intentionally not optimized, to show a
clear performance difference.

To run in preemptive mode, compile as normal.
To run in cooperative mode, compile with the following flags defined:
	KERNEL_MT_PURE_COOPERATIVE
	PROC_SW_TEST_MT_COOPERATIVE
To run in mixed mode, compile with the following flags defined:
	PROC_SW_TEST_MT_COOPERATIVE

App 1 repeatedly prints a short message.

App 2 repeatedly prints a long (~400 byte) message.

App 3 repeatedly prints a short message, and toggles the FRDM-K64 LED color
based on the value of a floating point operation.

To see performance changes, try changing the baud rate of UART0.  A very
slow baud shows the importane of using preemptive MT - app 3 is serviced
fairly while app 1 and app 2 block, waiting for the serial prints to
finish.  This effect is exacerbated in pure cooperative mode.