Context Switching, Preemption, SysTick timer, etc. (#13)
* Fixed up UNIX interfaces for upcoming IO changes.

 + uart_write no longer returns the available buffer space with zero
 bytes as an argument.
 + file table now includes a lock entry

These changes reflect a move towards a multi-process, shared-IO-context
style OS paradigm.  I don't really know what I'm doing, but imagine that
you have an application-level C program that consumes the entire
processor, but that application is multi-threaded: that's where we're
going.

* Force full linkage of asm.  Disable Extended Stack

 + Extended Stack Frames (on exception entry) are disabled until a
 process uses the floating point unit. (this is not well tested yet)
 + ASM code is now fully-linked (-Wl,--whole-archive). This is to
 prevent assembly language ISRs from being discarded by the linker.

* Streaming IO locks are now marshalled via SVC.

Service calls (SVC) are now used to perform read and write operations.
This has the benefit of implicitly disabling preemption during a read or
write operation, which causes occasional data corruption when there is
high process switch load and multiple processes are doing IO on the same
file.

Note that making PendSV a higher interrupt priority than SVC will break
this.

* Scheduler nearly complete and refactored.

Unfortunately, in order to do this neatly, a great deal of kernel
structure needed to be added.  Some of it is unnecessary, but worthwhile
in the long run, and ultimately makes the code more segmented and
structured.

 + The kernel now has "ports" - there is only the arm/cm4 port,
 currently.
 + Generic algorithms have their own files/modules at the kernel level,
 but their entry points are defined in the port.
 + The NVIC and Core priorities have been figured out.  Bug reports
 coming in hot to move everything to use NVIC_SetPriority due to
 MSB-aligned implementation (thanks, ARM.)

 - Kernel memory is a hot mess. There is a single file which allocates
 the process table (but not the file table, too many changes at once)
 - The drivers still allocate their own memory, which makes the previous
 bullet a design flaw.

More to come soon.  Switching on the SysTick is up next.  Ideally,
figuring out how to ensure SVC/PendSV do not preempt interrupts would be
nice.

* Added a SysTick timer driver.

 + It will try to use the 10 ms calibration value if it exists, but I
 can't test because the board I have doesn't have it

* Context Switching, System Calls, Preemption, oh my

This is the mega commit, the one no one wanted, but secretly needed.
 + Process-private stacks, context switching are now automatically
 managed by the kernel.
 + Preemptive and Cooperative multitasking modes are available.
 Processes which choose not to cooperate will be (forcibly) preempted by
 default.  To undo this, define KERNEL_MT_PURE_COOPERATIVE.
 + Added a file number for uart0.  It's a hack, but there is not time to
 do real files.
 + Kernel now initializes uart0, scheduler, and SysTick timer.
 + Entry point renamed to "init_main()" since only initialization of
 other processes should be performed there.  The process switcher will
 handle the rest.