Converted to hierarchical build definitions.
All that is necessary for a "blank" build is included - no programs are
run, and the kernel deadloops after initializing devices.

Note, all of this is not really production-quality OS ready.  The target
system still really only works with MK64F12, and the kernel_main does
not break out device initialization for many ARM registers.

Important changes:
 + programs should specify what devices they need in order to run.  This
 is done in the build description under `src/programs/meson.build`. See
 demo branch for example.
   - there is no way to specify which peripheral, currently. specifying
   `peripherals_required += ['uart']` enables an unspecified uart.
 + Specifying required devices/peripherals will enable them in other
 subdirectories.  The driver subdir will specify the appropriate flags
 to make that device appear in the build. `peripherals_required += ['uart']`
 will implicitly define `KINETIS_USE_UART`.
 + Programs with multiple tasks should follow this pattern for enabling
 those tasks in `init_main`.