Added XPC Relay on UART0 as a process.
The XPC Relay is initialized on UART0 at system startup.
Processes which need to communicate via XPC should first call
`xpc_yield_until_init()` to force them to yield until the relay is
ready.

Other processes writing to UART0 directly should avoid this behavior, as
it will interleave the XPC stream with raw data, which may be
interpreted as XPC messages.  Instead, reserving a "from" channel for
that process' output is advised.