Added a mutual-exclusion lock system call. (#25)
* Added a mutual-exclusion lock system call.

This is needed because mutexes need to be atomic, and we have no other
facilities for making that guarantee.

This feature was added to support the xpc relay, and the need to mutex
lock the relay's context.  I was hoping to not have to do this, but
alas, I forgot about the atomicity constraint of mutex locking. How
foolish.

* Fixed re-locking denial bug.